Move PHPUtil, PythonUtil etc. out of Slice
