+++ cpp/src/slice2php/PHPUtil.h
