+++ cpp/src/slice2py/Main.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 //
 
-#include <Slice/PythonUtil.h>
+#include "PythonUtil.h"
 #include <Slice/Util.h>
 #include <IceUtil/ConsoleUtil.h>
