+++ cpp/src/slice2py/PythonUtil.h
