+++ cpp/src/slice2rb/Ruby.cpp
@@ -10,7 +10,7 @@
 #include <IceUtil/ConsoleUtil.h>
 #include <Slice/Preprocessor.h>
 #include <Slice/FileTracker.h>
-#include <Slice/RubyUtil.h>
+#include "RubyUtil.h"
 #include <Slice/Util.h>
 
 #include <string.h>
