+++ cpp/src/slice2rb/RubyUtil.h
