+++ php/src/Operation.cpp
@@ -7,7 +7,7 @@
 #include <Proxy.h>
 #include <Types.h>
 #include <Util.h>
-#include <Slice/PHPUtil.h>
+#include "PHPUtil.h"
 
 using namespace std;
 using namespace IcePHP;
