+++ python/modules/IcePy/Operation.cpp
@@ -20,7 +20,7 @@
 #include <Ice/Properties.h>
 #include <Ice/Proxy.h>
 #include <IceUtil/Time.h>
-#include <Slice/PythonUtil.h>
+#include "PythonUtil.h"
 
 using namespace std;
 using namespace IcePy;
