+++ ruby/src/IceRuby/Slice.cpp
@@ -5,7 +5,7 @@
 #include <Slice.h>
 #include <Util.h>
 #include <Slice/Preprocessor.h>
-#include <Slice/RubyUtil.h>
+#include "RubyUtil.h"
 #include <Slice/Util.h>
 #include <IceUtil/Options.h>
